+++ src/main/java/net/finmath/information/package-info.java
@@ -2,7 +2,5 @@
  * Provides information about the library (e.g. the version and build number) and runtime.
  * 
  * @author Christian Fries
- * 
- * @author Christian Fries
  */
 package net.finmath.information;