+++ src/main/java/net/finmath/marketdata/model/curves/locallinearregression/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Provided classes implementing the local linear regression method, see see https://ssrn.com/abstract=3073942
+ * 
+ * @author Christian Fries
+ * @author Moritz Scherrmann
+ */
+package net.finmath.marketdata.model.curves.locallinearregression;