+++ src/main/java6/net/finmath/functions/SABRModel.java
@@ -18,7 +18,7 @@ public class SABRModel {
 	/**
 	 * 
 	 */
-	public SABRModel() {
+	private SABRModel() {
 		// TODO Auto-generated constructor stub
 	}