+++ src/main/java6/net/finmath/montecarlo/MonteCarloSimulationInterface.java
@@ -8,6 +8,7 @@ package net.finmath.montecarlo;
 import java.util.Map;
 
 import net.finmath.exception.CalculationException;
+import net.finmath.modelling.ModelInterface;
 import net.finmath.stochastic.RandomVariableInterface;
 import net.finmath.time.TimeDiscretizationInterface;
 
@@ -18,7 +19,7 @@ import net.finmath.time.TimeDiscretizationInterface;
  * 
  * @author Christian Fries
  */
-public interface MonteCarloSimulationInterface {
+public interface MonteCarloSimulationInterface extends ModelInterface {
 	
 	/**
 	 * Returns the numberOfPaths.