+++ src/main/java6/net/finmath/randomnumbers/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Wrapper classes to random number generators.
+ * 
+ * @author Christian Fries
+ */
+package net.finmath.randomnumbers;