+++ src/main/java6/net/finmath/stochastic/package-info.java
@@ -2,8 +2,6 @@
  * Interfaces specifying operations on random variables.
  * 
  * @author Christian Fries
- * 
- * @author Christian Fries
  */
 package net.finmath.stochastic;