+++ src/main/java6/net/finmath/time/businessdaycalendar/BusinessdayCalendarExcludingTARGETHolidays.java
@@ -24,13 +24,13 @@ public class BusinessdayCalendarExcludingTARGETHolidays extends BusinessdayCalen
 	private final BusinessdayCalendarInterface weekdayCalendar = new BusinessdayCalendarExcludingWeekends();
 	
 	/**
-	 * Create business day calendar.
+	 * Create TARGET business day calendar.
 	 */
 	public BusinessdayCalendarExcludingTARGETHolidays() {
 	}
 
 	/**
-	 * Create business day calendar using a given business day calendar as basis.
+	 * Create TARGET business day calendar using a given business day calendar as basis.
 	 * 
 	 * @param baseCalendar Calendar of business days.
 	 */
@@ -64,10 +64,10 @@ public class BusinessdayCalendarExcludingTARGETHolidays extends BusinessdayCalen
 	 * @return True, if date is easter sunday.
 	 */
 	public static boolean isEasterSunday(LocalDate date) {
-	    int Y = date.getYear();
-	    int a = Y % 19;
-	    int b = Y / 100;
-	    int c = Y % 100;
+	    int y = date.getYear();
+	    int a = y % 19;
+	    int b = y / 100;
+	    int c = y % 100;
 	    int d = b / 4;
 	    int e = b % 4;
 	    int f = (b + 8) / 25;
@@ -75,10 +75,10 @@ public class BusinessdayCalendarExcludingTARGETHolidays extends BusinessdayCalen
 	    int h = (19 * a + b - d - g + 15) % 30;
 	    int i = c / 4;
 	    int k = c % 4;
-	    int L = (32 + 2 * e + 2 * i - h - k) % 7;
-	    int m = (a + 11 * h + 22 * L) / 451;
-	    int easterSundayMonth	= (h + L - 7 * m + 114) / 31;
-	    int easterSundayDay		= ((h + L - 7 * m + 114) % 31) + 1;
+	    int l = (32 + 2 * e + 2 * i - h - k) % 7;
+	    int m = (a + 11 * h + 22 * l) / 451;
+	    int easterSundayMonth	= (h + l - 7 * m + 114) / 31;
+	    int easterSundayDay		= ((h + l - 7 * m + 114) % 31) + 1;
 
 		int month = date.getMonthValue();
 		int day = date.getDayOfMonth();