+++ src/main/java6/net/finmath/time/businessdaycalendar/package-info.java
@@ -2,8 +2,6 @@
  * Provides business day calendars, e.g., as used in date roll conventions.
  * 
  * @author Christian Fries
- * 
- * @author Christian Fries
  */
 package net.finmath.time.businessdaycalendar;