+++ src/main/java6/net/finmath/time/daycount/package-info.java
@@ -2,8 +2,6 @@
  * Provides various day count conventions, e.g., as used in the definition of coupon payments of interest rate products.
  * 
  * @author Christian Fries
- * 
- * @author Christian Fries
  */
 package net.finmath.time.daycount;